Move combat to its own file
