--- src/simulation.log
@@ -1,3 +1,0 @@
-Simulated 1000 games in 6.97s: 143 games/sec
-Simulated 1000 games in 7.61s: 131 games/sec
-Simulated 1000 games in 7.72s: 129 games/sec
